[I25-256] feat : import 경로 수정 및 기타 리팩토링
