+++ src/app/providers.tsx
@@ -2,7 +2,7 @@
 
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { useState } from 'react';
-import { ModalProvider, Modal } from '@components/modal';
+import { ModalProvider, Modal } from '@/app/components/modal';
 
 export default function Providers({ children }: { children: React.ReactNode }) {
   const [queryClient] = useState(
