+++ src/services/project/getProjects.server.ts
@@ -4,8 +4,8 @@ import { createClient } from '@/services/supabase/server';
 import { CardProps } from '@/app/components/card/project/ProjectCard';
 import { Tables } from '@/services/supabase/database.types';
 import { convertFromDatabaseImageURL } from '@/services/supabase/imageHostConverter';
-import type { ProjectContributor, ProjectOwnerProfile } from './types';
-import { createAuthorsFromProject } from './utils';
+import type { ProjectContributor, ProjectOwnerProfile } from '@/services/project/types';
+import { createAuthorsFromProject } from '@/services/project/utils';
 
 export type ProjectType = Pick<
   Tables<'projects'>,
