+++ src/services/project/utils.ts
@@ -1,4 +1,4 @@
-import type { ProjectContributor, ProjectOwnerProfile } from './types';
+import type { ProjectContributor } from '@/services/project/types';
 import { convertFromDatabaseImageURL } from '@/services/supabase/imageHostConverter';
 
 export type ProjectWithContributors = {
@@ -9,12 +9,14 @@ export function createAuthorsFromProject(
   project: ProjectWithContributors,
 ): Array<{ name?: string; profileImage: string }> {
   return (
-    project.project_contributors?.map((contributor) => ({
-      name: contributor.profile.profile_name,
-      profileImage: convertFromDatabaseImageURL(
-        contributor.profile.profile_image,
-      ),
-    })) || []
+    project.project_contributors
+      ?.filter((contributor) => contributor.profile !== null)
+      .map((contributor) => ({
+        name: contributor.profile!.profile_name,
+        profileImage: convertFromDatabaseImageURL(
+          contributor.profile!.profile_image,
+        ),
+      })) || []
   );
 }
